add default config test
