--- examples/wasm/config/v1alpha1/BUILD
@@ -1,10 +1,0 @@
-cc_proto_library(
-  name = "config_cc_proto",
-  visibility = ["//examples/wasm:__pkg__",],
-  deps = ["config_proto"],
-)
-
-proto_library(
-  name = "config_proto",
-  srcs = ["config.proto"],
-)
