--- examples/wasm/config/v1alpha1/config.proto
@@ -1,8 +1,0 @@
-syntax = "proto3";
-
-package examples.wasm.config.v1apha1;
-
-message PluginConfig {
-    string name = 1;
-    string value = 2;
-}
